+++ test/integration/client/restore_sent_from_trash_test.dart
@@ -1,4 +1,5 @@
-import 'package:enough_mail_plus/enough_mail.dart';
+import 'package:ndk/shared/nips/nip19/nip19.dart';
+import 'package:nostr_mail/nostr_mail.dart';
 import 'package:test/test.dart';
 
 import '../../helpers/test_user.dart';
@@ -23,9 +24,10 @@ void main() {
 
     await sender.client.send(
       to: [
-        MailAddress(
-          null,
-          'npub1krtvaf2gw0ukuvgxvf7kxjz8s3zd6agfk87cnpcdha0s8xuscj2qly5eac@nostr',
+        NostrRecipient.fromPubkey(
+          Nip19.decode(
+            'npub1krtvaf2gw0ukuvgxvf7kxjz8s3zd6agfk87cnpcdha0s8xuscj2qly5eac',
+          ),
         ),
       ],
       subject: 'Restore sent message',